requirements.txt
flask==3.0.0
flask-cors==4.0.0
pypdf==4.0.0
reportlab==4.1.0
Pillow==10.0.0
pdf2image==1.16.3
python-bidi==0.4.2
gunicorn==21.2.0
